+++ src/main/java/titarenko/test2/repo/jdbc/JdbcContactRepoImpl.java
@@ -42,7 +42,7 @@ public class JdbcContactRepoImpl extends JdbcDaoSupport implements ContactRepo {
     }
 
     @Override
-    public List<Contact> getContactsPart(Long from, Long to) {
+    public List<Contact> getContactsPart(Integer from, Integer to) {
         return getJdbcTemplate().query("select * FROM contacts where id > ? and id < ?", ROW_MAPPER, from, to);
     }
